Corrected the Animal class constructor from int to double for the totalWeight parameter. The same correction is made in the BirdsCattle.cs, Chicken.cs, Duck.cs and Turkey.cs

+++ TheButcherCalulatorConsole/TheButcherCalculator/Animal.cs
@@ -36,7 +36,7 @@ namespace TheButcherCalculator
             }
         }
         //constructors
-        protected Animal(string kind, int totalWeight)
+        protected Animal(string kind, double totalWeight)
         {
             this.Kind = kind;
             this.TotalWeight = totalWeight;
